requirements.txt
typing-extensions
langchain-openai
langchain-experimental
pandas
yfinance
praw
feedparser
stockstats
eodhd
langgraph
chromadb
setuptools
backtrader
akshare
tushare
finnhub-python
parsel
requests
tqdm
pytz
redis
chainlit
rich
questionary
# For OpenAI LLM support
openai
# For local LLM support (Llama models)
llama-cpp-python
# For local embedding support
sentence-transformers
